+++ requirements.txt
@@ -5,8 +5,8 @@ Flask~=2.0.2
 Flask-Seasurf~=0.3.1
 Flask-SQLAlchemy~=2.5.1
 flask-talisman~=0.8.1
-grpcio==1.42.0
-grpcio-tools==1.42.0
+grpcio==1.43.0
+grpcio-tools==1.43.0
 psycopg2~=2.9.2
 python-dotenv~=0.19.2
 requests~=2.26.0
